# avoid allocating new label Builder in loop