@pracucci 2021-02-23
Thanks @nicolai86. To better compare benchmarks, please do the following:
1. Run each benchmark with `-count=5` and store the output to a file (eg. `before.txt` and `after.txt`)
2. Install [`benchstat`](https://pkg.go.dev/golang.org/x/perf/cmd/benchstat) and run `benchstat before.txt after.txt`

Benchstat will include statistical significance in the comparison.