@nicolai86 2021-02-23
```
❯ benchstat before.txt after.txt
name                   old time/op  new time/op  delta
GetLabels_SmallSet-8   3.67µs ± 3%  2.60µs ± 2%  -29.12%  (p=0.008 n=5+5)
GetLabels_MediumSet-8  1.66ms ±10%  0.79ms ± 8%  -52.65%  (p=0.008 n=5+5)
```