@pracucci 2021-02-24
> benchstat before.txt after.txt

@nicolai86 What you provided is fine, thanks! For the next time, please also run tests with `-benchmem` so the memory allocations / allocated is compared as well.

Could you also sign your commit (DCO), otherwise we can't merge?