@nicolai86 2021-02-24
thanks for the hand holding @pracucci .

signed off. also, here's an updated benchstat:

```
❯ benchstat before.txt after.txt
name                   old time/op    new time/op    delta
GetLabels_SmallSet-8     3.73µs ± 4%    2.75µs ± 7%  -26.31%  (p=0.000 n=10+10)
GetLabels_MediumSet-8    1.83ms ±16%    0.93ms ±15%  -49.23%  (p=0.000 n=10+10)

name                   old alloc/op   new alloc/op   delta
GetLabels_SmallSet-8     2.40kB ± 0%    1.20kB ± 0%  -50.00%  (p=0.000 n=10+10)
GetLabels_MediumSet-8     889kB ± 0%     169kB ± 0%  -81.00%  (p=0.000 n=9+6)

name                   old allocs/op  new allocs/op  delta
GetLabels_SmallSet-8       30.0 ± 0%      20.0 ± 0%  -33.33%  (p=0.000 n=10+10)
GetLabels_MediumSet-8     12.0k ± 0%      6.0k ± 0%  -49.96%  (p=0.000 n=10+10)
```