@bboreham 2021-02-25
Would it be even better to put off building until we are sure this metric matches the filter?
On the assumption that the most common case it that it does not match.