@nicolai86 2021-02-25
with my naive implementation - introducing a variable, creating and resetting only once `lbls` is used - it improves the case without match ~15%. 

However it adds one allocation again, which makes the case with matches slightly slower to what it is with the default allocation right now.