@bboreham 2021-02-25
Don’t those numbers depend on how many labels are in each metric?