@nicolai86 2021-02-25
It should - the number is just for the small set benchmark I added as part of this PR.